composite the app to mafia app
